Added sign up page and started password hashing
Created a new page solely for adding new users. When the login credentials are entered, they're to be comited to the main database and the user is taken back to the login page. I also started work on password hashing and would like you guys to take a look and see if you can add to it and test it to get it running.